Crate domain_map

This module provides a simple datastructure that can store values associated with a domain name style key. Wildcard keys are supported.

Structs

DomainMap

A DomainMap is conceptually similar to a HashMap. The differences are that the keys are always domain name strings like “example.com” and that a lookup that doesn’t have an exact match in the map is allowed to resolve through a wildcard entry, such as “.example.com”, if one has been inserted. A lookup for “example.com” will not match the wildcard “.example.com” because it has fewer segments than the wildcard entry.

Functions

register